refactor(db): audit fixes — sentinel errors, user_version, integer money, schema constraints, test split
- goal.rs, task.rs: eliminate QueryReturnedNoRows sentinel hack; return
  row counts from closures and check outside
- handle.rs: add PRAGMA user_version migration tracking
- types.rs: change budget_usd and checkpoint values from f64 to i64 (cents)
- migrations/001_initial.sql: add CHECK(slice_execution IN (0,1)),
  idx_budget_checkpoints_goal_id, change budget_usd/limit_value/used_value to INTEGER
- transaction.rs: add #[must_use], improve Drop documentation
- task.rs: validate goal_id consistency in create_batch, add get_by_id
- event.rs: add event_id secondary sort for deterministic ordering
- tests.rs: split into tests/ directory module (mod.rs, crud.rs,
  transaction.rs, concurrent.rs, misc.rs) to comply with 400-line limit

+++ src/runtime/db/types.rs
@@ -13,7 +13,7 @@ pub struct GoalRecord {
     pub max_agents: Option<i32>,
     pub budget_time_secs: Option<i64>,
     pub budget_tokens: Option<i64>,
-    pub budget_usd: Option<f64>,
+    pub budget_usd: Option<i64>, // cents
     pub created_at: i64,
     pub updated_at: i64,
     pub controller_pid: Option<i32>,
@@ -67,8 +67,8 @@ pub struct BudgetCheckpoint {
     pub checkpoint_id: Option<i64>,
     pub goal_id: String,
     pub kind: String,
-    pub limit_value: Option<f64>,
-    pub used_value: Option<f64>,
+    pub limit_value: Option<i64>,
+    pub used_value: Option<i64>,
     pub created_at: i64,
 }
